Add files via upload
Is there a particular reason why my code appears in the way I want it to for firefox and chrome but is wrecked in IE? I'm aware of the differences between the browsers but prior to me getting to the current result, it didn't look this cluttered. I'm confused; it may be my code and if it is let me know what I can do to fix it. Thanks.